+++ .gitignore
@@ -1,2 +1,3 @@
 .DS_STORE
 .env
+node_modules/
