+++ package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node api/app.js"
+    "start": "node app.js"
   },
   "repository": {
     "type": "git",
